Disable HeaderView singleton checks for js-tests
Some updates to the pages framework on master had made the `HeaderView` a
singleton object. This caused conflicts when running the tests, since
a `HeaderView` is instantiated by the test page, and then each of the
individual pages would create additional ones when they were being
tested.

The tests for the individual pages are pretty isolated, and access the
`HeaderView` through their stored attributes instead of as a singleton.
It's therefore safe to have multiple ones when we're in the context of
the test runner. This change adds spies in the right place to suppress
the warnings.

Testing Done:
Ran the JS test suite and saw that a bunch of warnings about duplicate
singleton objects no longer showed up.

Reviewed at https://reviews.reviewboard.org/r/11412/