Add formal support for spinners on buttons.
This makes use of the new spinner support in Djblets, adding the ability
to show a spinner on a button using styling appropriate for the state
the button is in.

It also fixes up some styling for legacy `.btn` elements.

Testing Done:
Tested this with the Enable/Disable buttons on the new Extensions page.
Verified the spinner was styled correctly for all states.

Reviewed at https://reviews.reviewboard.org/r/11046/