Fix a crash on Permission Denied in the Review General Comment API.
When trying to POST to the Review General Comment API without permission
to modify the review, the API would pass to the wrong variable to
`get_no_access_error()`, causing an HTTP 500.

This is a simple change that fixes the variable and adds a unit test to
ensure the fix does not regress.

Testing Done:
Unit tests pass.

Bugs closed: 4850

Reviewed at https://reviews.reviewboard.org/r/11052/