Fix the "Public Read Access" indicator in the admin UI.
The "Public Read Access" indicator shown in the sidebar in the
administration UI was being based on the wrong siteconfig setting,
causing it to show a disabled state when public access was enabled. This
was purely cosmetic but confusing.

This change fixes up the template to access the correct variable.

Testing Done:
Verified that the indicator reflected the correct state when public
access was enabled and disabled.

Reviewed at https://reviews.reviewboard.org/r/11060/