Update the Storage Settings page for modern subform support.
The Storage Settings page has been updated to work with the subform
support introduced in Review Board 4.0. This gives us a maintainable way
of allowing authentication forms to be added to the page and configured
without triggering any browser-side validation issues for unselected
subforms.

In order for this to work, the existing settings form had to be split up
into subforms. Ideally, we'd have pluggable backends we'd load settings
forms from, but this is at least a step toward that.

Testing Done:
Tested saving the search form with each storage backend subform selection.
Verified that only the selected subforms received any client-side or
server-side validation, and that settings loaded and saved.

Reviewed at https://reviews.reviewboard.org/r/11065/