Tweak the UI and code for the Avatar Settings page.
This makes a few small but helpful improvements to the Avatar Settings
page. The major change is that the side-by-side selection widget for the
avatar backends (which doesn't work well on mobile and is largely being
eliminated from the UI) has been replaced by a set of checkboxes for
each service. This is far more clear than the old UI, and works great on
mobile and desktop.

The ordering of fields has changed, along with their labels. We now have
"Enable avatars" first, followed by a list of services, and followed by
the default service. This is a more logical order.

There's new help text that explains what happens if a service isn't
enabled, and displays our recommendation to keep avatars enabled.

And last, but not least, the page title has changed to "Avatar
Settings", and the sidebar entry to "Avatars", for consistency with the
rest of the settings.

Testing Done:
Tested the page on desktop and mobile, making sure all options loaded
and persisted.

Reviewed at https://reviews.reviewboard.org/r/11072/