Fix dirty state calculations for inline editors with clipboard events.
When clipboard events are triggered via the UI rather than the keyboard
(for example, right-click menu options or middle-click paste), it
wouldn't trigger an update of the dirty state, resulting in the editor
resetting when the "OK" button was clicked. This is trivially fixed by
updating the dirty state when we get `cut` or `paste` events.

Testing Done:
Used the right-click menu to edit a field with both cut and paste
operations. Saw that the contents were correctly saved when clicking
"OK" with both markdown and plain-text modes.

Reviewed at https://reviews.reviewboard.org/r/11109/