Update a legacy import in rb-site to use a compatibility module.
`rb-site upgrade` attempts to upgrade very ancient cache backend
settings when found. To do this, it needs to import
`parse_backend_uri()`, which Django no longer provides. However, Djblets
has a compatibility import for this purpose.

This is a trivial change that updates our import to use the
compatibility import.