Fix a crash when Paramiko has trouble with loading a host key.
It's possible for Paramiko to raise an exception if it can't load a host
key that's in a format it doesn't understand. This breaks any attempts
at loading host keys. This is most likely to happen on a developer
setup.

We now check for this and just skip the key.

Testing Done:
Encountered this locally and saw it affect unit test runs and some
repository interaction. After the patch, everything worked as expected.

Reviewed at https://reviews.reviewboard.org/r/11110/