Don't crash if a bad LocalSite ID is posted in a LocalSite-aware form.
`LocalSiteAwareModelFormMixin` uses the `LocalSite` ID passed when
posting form data and validates all other relation fields against that.
An assumption is made that we can perform a query for a `LocalSite`
given that ID, and on some databases that works fine, but on others it
will raise an exception if the ID isn't numeric.

We now check for this ahead of time and ignore the value if provided,
avoiding a confusing crash.

Testing Done:
Unit tests pass.

Verified the crash occurred without this fix but was fixed with it.

Reviewed at https://reviews.reviewboard.org/r/11111/