Update the docs to cover all directories needing ownership updates.
We have documentation that tells people to `chown` a few directories,
but it wasn't comprehensive. While we do provide comprehensive
information when we detect bad ownership or permissions, it helps to not
be misleading in the documentation.

The docs are now updated to cover all relevant directories up-front.

Testing Done:
Built the docs and checked the results.

Reviewed at https://reviews.reviewboard.org/r/11112/