Improve the error message when encountering an unverified certificate.
The error message saying that repository communication encountered an
unverified certificate didn't really do enough to help diagnose the
problem. All the user was left with was an idea that something went
wrong, but not really what that was.

This expands on it a bit, placing additional information within the
error message to identify the hostname and the fingerprint, so users can
confirm whether they match expectations.

Testing Done:
Tested along with an upcoming change to Perforce. Verified the error
response when encountering an unverified SSL certificate.

Configured a Perforce repository with a self-signed SSL certificate
and checked the prompt to make sure it was using the repository form's
standard text.

Reviewed at https://reviews.reviewboard.org/r/11125/