Improve Perforce SSL fingerprint parsing and error reporting.
The Perforce SSL error handling assumed some things about the placement
of the fingerprint in the error message. This wasn't always correct,
though. While this hasn't really been an issue in the past, practically
speaking, it is now that we have improved unverified SSL certificate
error.

We now look for the fingerprint using a regex. We don't care where it is
in the string, so long as it's there somewhere.

Both that and the P4PORT get passed into the `Certificate` used for the
error message, which will then be displayed to the user.

Testing Done:
Triggered the error with two kinds of SSL cert failures (one triggered
by no existing stored fingerprints for a server, one triggered by a new
fingerprint when an existing one was stored). Verified the error output
was correct in both cases.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11126/