Enhance and fix up the setup experience in prepare-dev.py.
This modernizes the way we request input and display output on
`prepare-dev.py`, utilizing more of rb-site's `ConsoleUI`. This is a
better, more consistent way of displaying information and prompting for
information.

The text has been updated to be more user-friendly, to offer
guidance for when things go wrong, and to provide next steps when the
environment is prepared.

Testing Done:
Ran through `prepare-dev.py`. Saw that I got all the nice text UI
enhancements and string changes.

Reviewed at https://reviews.reviewboard.org/r/11162/