Add pipelined static media compilation to prepare-dev.py.
When setting up a new tree, a common annoyance is that the first page
load has to compile all the referenced static media. This may appear to
be a problem with the server loading.

To improve the initial experience, we now compile all the static media
bundles up-front through Django Pipeline. This can take some time, but
no longer than it would take if trying to load a page and doing it then.
We also show the progress as the compilations proceed, giving a sense of
how much longer the process will take.

Testing Done:
Ran `prepare-dev.py` on a new tree. Saw it compile all the static media
files, and verified their presence in the `htdocs/static/` directory.

Reviewed at https://reviews.reviewboard.org/r/11164/