Add color output and visual hints to rb-site and prepare-dev.py.
Both rb-site and prepare-dev.py deal with a lot of text, and it can kind
of blend together, particularly as operations are being performed
(database synchronization, for instance).

To help distinguish parts of the setup, we now display color output for
headers, prompts, errors, notices, and success indicators.

Headers also have a visual separator, making it very clear what section
some content is in. This replaces the simple asterisk that previously
proceeded a section title.

Color is only enabled if the terminal supports it (piggybacking on the
color support for Django management commands). `--no-color` can also be
provided to disable color.

Text content will also now fit the width of the terminal when run on
Python 3.x, which helps with cutting off content prematurely. This makes
use of `shutil.get_terminal_size()`, which is a standard way of returning
terminal dimensions.

Testing Done:
Created a new site. Checked that all color output was as expected.

Set up a new dev tree. Verified the color output there as well.

Reviewed at https://reviews.reviewboard.org/r/11165/