Fix GitHub auth token bug using wrong HTTP header capitalization in Python2.7
When trying to add a newly configured repo, it complained about the
GitHub Auth Token missing required permissions.

The underlying issue was that the wrong capitalization was used when
getting the OAuth scopes header for Python2.7.

Testing Done:
Did the following with all relevant versions of Python3 and 2:

* Ran unit tests and they passed

* Manually added a repo with with a personal access token

Reviewed at https://reviews.reviewboard.org/r/11173/