Fix wrapping on newly-created review replies.
The client-side template for the review reply editor was creating the
`<pre>` tag for the reply contents missing the `comment-text` class,
which didn't cause many obvious problems except that the text would not
wrap correctly. A reload would fix this, since then the reply element
was created on the backend with the correct class list. This change
fixes up the template to add all necessary classes.

This only appeared on reviews associated with status updates, because
there was a separate rule setting `white-space` on pre tags within a
`.review` element. These should probably all be cleaned up at some point
when we move the review entries over to the new CSS component standard.

Testing Done:
Created a draft reply to a Review Bot comment and saw that it now
wrapped correctly.

Reviewed at https://reviews.reviewboard.org/r/11221/