Always send Authorization header on authenticated webhooks
The problem with `HTTPBasicAuthHandler` is that it will first attempt to
make an unauthenticated request, and will retry with authentication only
if a 401 response was seen. However, some servers respond with 403 if
the right credentials are not supplied right away. This change alters
the webhook notification to unconditionally send the HTTP Basic header
if credentials are present in the webhook URL.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11193/