Fix My Privacy Rights showing empty label.
The "My Privacy Rights" page in the user account settings was showing an
orphaned `:*`. This was for the field label for the consent items, but that
form field didn't have a label defined. The cause here was that the Review
Board form field template was checking if the form had a `fields_no_label`
attribute, but that was deprecated and removed from Djblets a long time ago.
The correct fix is to check if the field has a label defined.

Testing Done:
Manual testing by refreshing the My Privacy Rights page and checking that the
empty label is no longer visible. Other labels on the account forms that are
supposed to be there have been verified to be visible.

Doing a search of `fields_no_label` on all the reviewboard repositories show
no usage of that particular variable.

Bugs closed: 4882

Reviewed at https://reviews.reviewboard.org/r/11180/