Fix unit tests that are sensitive to timezone changes.
We had some user profile unit tests that assumed US/Pacific would map to
-0700. The Daylight Savings switch broke these tests, not surprisingly.

This change calculates what the correct timezone offset should be, based
on whether we're in Daylight Savings or not, and then checks for that in
the calculated payload.

Important caveat: This may break in the future if California separates
from the continent and floats into the ocean.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11263/