Backport RemovedInReviewBoard50Warning to 3.0.x
We'll be making a small API change to allow SCMTool filename
normalization to make use of the extra_data field. We'll want this to be
3.0.19 and newer, but won't actually remove the deprecation of the older
APIs until 5.0. The release-4.0.x branch has the
`RemovedInReviewBoard50Warning` and the `warn` helper method, but 3.0.x
didn't. This change backports it.

Testing Done:
Used in conjunction with other changes.

Reviewed at https://reviews.reviewboard.org/r/11272/