Pass through FileDiff extra_data into filename normalization.
While most version control systems are relatively sane and treat a
filename as a filename, there are a few that don't. We first encountered
this with CVS, which loves to add a `,v` to the end of filenames
depending on the method of access. To address this, we added a
normalization method that SCMTools could override.

Impending work on improving support for ClearCase requires even fancier
normalization, and one that requires being able to use some data about
the repository. This change adds the FileDiff extra_data as an argument
to the normalization method.

This is a soft API break. Existing SCMTools which provide a normalize
method that does not take any additional arguments will continue to
function, but emit a deprecation warning.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11282/